requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
openai==1.12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mcp==1.24.0
pydantic==2.5.0
typing-extensions==4.9.0
anyio==3.7.1
starlette==0.27.0
httpx==0.25.2
click==8.1.7
google-genai>=0.3.0